Refactor to app factory
