+++ tests/test_cors_security.py
@@ -8,11 +8,11 @@ def load_app(monkeypatch, cors_value):
     monkeypatch.setenv('TWILIO_ACCOUNT_SID', 'dummy')
     monkeypatch.setenv('TWILIO_AUTH_TOKEN', 'dummy')
     monkeypatch.setenv('TWILIO_WHATSAPP_FROM', 'dummy')
-    for module in ["main", "app.config"]:
+    for module in ["wsgi", "app", "app.config"]:
         if module in sys.modules:
             del sys.modules[module]
-    main = importlib.import_module("main")
-    return main.app
+    entry = importlib.import_module("wsgi")
+    return entry.app
 
 
 def test_cors_preflight_allows_whitelisted_origin(monkeypatch):
