+++ wsgi.py
@@ -1,1 +1,4 @@
-from main import app
+from app import create_app
+
+# WSGI entry point for production and tests
+app = create_app()
